+++ tests/testthat/test-metadata.R
@@ -39,9 +39,9 @@ test_that("can get plot metadata for a country", {
       "colour", "min", "max", "invert_scale") %in%
       names(metadata)))
   expect_true(all(unique(metadata$indicator) %in%
-                  c("art_coverage", "art_current", "receiving_art",
+                  c("art_coverage", "art_current", "art_current_residents",
                     "prevalence", "art_number",
-                    "incidence", "new_infections", "plhiv", "population",
+                    "incidence", "infections", "plhiv", "population",
                     "recent_infected", "viral_suppression_plhiv",
                     "anc_prevalence", "anc_art_coverage")))
 })
@@ -55,23 +55,26 @@ test_that("can get plot metadata for missing country with defaults", {
       "error_high_column", "indicator_column", "indicator_value", "name",
       "colour", "min", "max", "invert_scale") %in%
       names(metadata$result)))
-  expect_true(all(unique(metadata$result$indicator) %in%
-                  c("art_coverage", "art_current", "receiving_art",
-                    "prevalence", "art_number",
-                    "incidence", "new_infections", "plhiv", "population",
+  expect_setequal(metadata$result$indicator,
+                  c("art_coverage", "art_current", "art_current_residents",
+                    "prevalence", 
+                    "incidence", "infections", "plhiv", "population",
                     "recent_infected", "viral_suppression_plhiv",
                     "anc_prevalence", "anc_art_coverage",
-                    "anc_clients", "anc_plhiv", "anc_already_art")))
+                    "anc_clients", "anc_plhiv", "anc_already_art",
+                    "anc_art_new", "anc_known_pos", "anc_tested_pos",
+                    "anc_tested_neg"))
 })
 
 test_that("colour scales metadata is well formed", {
   scales <- naomi_read_csv(system_file("metadata", "colour_scales.csv"))
-  expect_true(all(scales$indicator %in%
-    c("art_coverage", "art_current", "receiving_art", "prevalence",
+  expect_setequal(scales$indicator,
+    c("art_coverage", "art_current", "art_current_residents", "prevalence",
       "viral_suppression_plhiv", "recent_infected",
-      "art_number", "plhiv", "incidence", "population", "new_infections",
+      "plhiv", "incidence", "population", "infections",
       "anc_prevalence", "anc_art_coverage", "anc_clients", "anc_plhiv",
-      "anc_already_art")))
+      "anc_already_art", "anc_art_new", "anc_known_pos", "anc_tested_pos",
+      "anc_tested_neg"))
   expect_equal(nrow(unique(scales[, c("iso3", "indicator")])), nrow(scales))
   expect_true(is.numeric(scales$min))
   expect_true(is.numeric(scales$max))
@@ -91,24 +94,25 @@ test_that("colour scales metadata is well formed", {
 
 test_that("metadata is well formed", {
   meta <- get_metadata()
-  expect_true(all(meta$indicator %in%
-                  c("art_coverage", "art_current", "prevalence",
+  expect_setequal(meta$indicator,
+                  c("art_coverage", "art_current", "art_current_residents", "prevalence",
                     "viral_suppression_plhiv", "recent_infected", "plhiv",
-                    "incidence", "art_number", "population", "incidence",
-                    "new_infections", "receiving_art", "anc_prevalence",
-                    "anc_art_coverage", "anc_clients", "anc_plhiv", "anc_already_art")))
+                    "population", "incidence", "infections", "anc_prevalence",
+                    "anc_art_coverage", "anc_clients", "anc_plhiv", "anc_already_art",
+                    "anc_art_new", "anc_known_pos", "anc_tested_pos", "anc_tested_neg"))
   expect_equal(nrow(unique(meta[, c("data_type", "plot_type", "indicator")])),
                nrow(meta))
   expect_true(all(meta$plot_type %in% c("choropleth", "barchart")))
   expect_true(all(meta$data_type %in% c("survey", "anc", "programme", "output")))
-  expect_true(all(meta$name %in%
+  expect_setequal(meta$name,
                   c("HIV prevalence", "ART coverage", "Viral load suppression",
                     "Proportion recently infected", "PLHIV", "Population",
-                      "New infections", "HIV incidence", "ART number (residents)",
+                    "New infections", "HIV incidence", "ART number (residents)",
                     "ART number (attending)", "ANC HIV prevalence",
                     "ANC prior ART coverage", "ANC clients",
                     "HIV positive ANC attendees",
-                    "ANC attendees already on ART")))
+                    "ANC attendees already on ART", "ART initiations at ANC",
+                    "ANC known positive", "ANC tested positive", "ANC tested negative"))
   expect_equal(
     colnames(meta),
     c("data_type", "plot_type", "indicator", "value_column", "error_low_column",